Update versions for: Catch2 and JSON libs
New version of Catch is a major, breaking change.
Catch2 stopped being used as header-only library.
Now it is distributed as static library.
This may cause a classic issue of lack of prebuilt libs in the package.
For example, at this time there is no package for M1.
In such case you need to build it locally:
```
conan install . --build=catch2
```